Update README.md
Built with a simple meaning of practicing my web development skills, after each discover and things I learnt I got more obssessed with the page and more ideas were appearing as the time passed. This is the biggest project I've done so far and made me learn a lot. 
The page includes: HTML, CSS, Javascript (jQuery and AJAX for backend communications) and JSON for frontend, PHP and MySQL for backend.

Features included:

Banlist and matchmaking:

1. Data fetching
2. Communication between CS1.6 Servers (ban system and match making system)
3. Ban list system shared among 10+ servers
4. Web Panel for ban players
5. Dashboard to watch Matchmaking histories both LIVE or Old matches
6. Statistics of each match with a lot of infos about players and messages
7. Comment system
8. Ranking system shared among all servers that uses the EMatch system
9. Player profile with options to customize the background and images with gifs (border, background, mini-background, player name color, avatar and nickname icon)
10. Servers list that updates the informations of the server (such as name, map, players, offline/online status and an option to connect)
11. A session of the website totally directed to the automix to explain and let players know what the system has.
    
Some common options:

1. Language toggle
2. Light/dark mode toggle
3. Visitor count
4. Reponsive
5. Christmas and New Year Themes avaliable
6. Page maintenance option which blocks any guests to join, only the logged manager can join
7. Page debug
8. A lot of manager options, they are: Ban/unban player through the site, post/delete/edit plugins posted on the download page, delete a match from the database on the dashboard and delete comments on the plugins.

On the page, there are also some plugins able to download for the community. The intention is to help then have more options to play with and easy to configure. The plugins are written in PAWN, made by Amxmodx Team. 
All the plugins on the list are saved with JSON so the search and loading are faster. Basic options for searching that automatically toggles the fetching when the user stops writting the input text, pagination, back to top button and sidebar.

Some pictures:

Download Page

Light:
![My animated logo](https://i.imgur.com/lODCLEr.png)
![My animated logo](https://i.imgur.com/pYxj0p0.png)
![My animated logo](https://i.imgur.com/TJf4IWS.png)

Dark:

![My animated logo](https://i.imgur.com/C281nWc.png)
![My animated logo](https://i.imgur.com/7NHb5PS.png)
![My animated logo](https://i.imgur.com/6Gclpha.png)

Dashboard
![My animated logo](https://imgur.com/N3me0NC.png)

Ranking
![My animated logo](https://i.imgur.com/CFH9Gfy.png)

Player Profile
![My animated logo](https://i.imgur.com/fv08L72.png)

Custom Options
![My animated logo](https://i.imgur.com/DfoTKHY.png)

Match Statistics
![My animated logo](https://i.imgur.com/kJjEsXN.png)
![My animated logo](https://i.imgur.com/hf9JSNu.png)

Servers List
![My animated logo](https://i.imgur.com/Q3WNTVk.png)

Ban List
![My animated logo](https://i.imgur.com/C1fMWx1.png)
![My animated logo](https://i.imgur.com/bdRIpjU.png)

Pug Main Page
![My animated logo](https://i.imgur.com/xr9yVZA.png)
![My animated logo](https://i.imgur.com/wPATny7.png)
![My animated logo](https://i.imgur.com/f4gb00D.png)
![My animated logo](https://i.imgur.com/9uHoznF.png)

Web Panel to manage ban from the servers
![My animated logo](https://i.imgur.com/EBNwcjX.png)
![My animated logo](https://i.imgur.com/4JBfrfW.png)
![My animated logo](https://i.imgur.com/gjiG7Fz.png)
![My animated logo](https://i.imgur.com/0NGzDhb.png)